ddl: migrate test-infra to testify for TestModifyColumn*
