+++ ddl/placement_sql_test.go
@@ -26,7 +26,6 @@ import (
 	mysql "github.com/pingcap/tidb/errno"
 	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/testkit"
-	"github.com/pingcap/tidb/util/testutil"
 	"github.com/stretchr/testify/require"
 )
 
@@ -316,13 +315,13 @@ func TestAlterDBPlacement(t *testing.T) {
 	// Test for information_schema.schemata
 	tk.MustQuery("SELECT CATALOG_NAME, SCHEMA_NAME, DEFAULT_CHARACTER_SET_NAME, DEFAULT_COLLATION_NAME, TIDB_PLACEMENT_POLICY_NAME FROM information_schema.schemata WHERE SCHEMA_NAME='TestAlterDB'").Check(testkit.Rows(`def TestAlterDB utf8mb4 utf8mb4_bin alter_x`))
 	// Test for Show Create Database
-	tk.MustQuery(`show create database TestAlterDB`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create database TestAlterDB`).Check(testkit.RowsWithSep("|",
 		"TestAlterDB CREATE DATABASE `TestAlterDB` /*!40100 DEFAULT CHARACTER SET utf8mb4 */ "+
 			"/*T![placement] PLACEMENT POLICY=`alter_x` */",
 	))
 	// Test for Alter Placement Policy affect table created.
 	tk.MustExec("create table t(a int);")
-	tk.MustQuery(`show create table t`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create table t`).Check(testkit.RowsWithSep("|",
 		"t CREATE TABLE `t` (\n"+
 			"  `a` int(11) DEFAULT NULL\n"+
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin "+
@@ -332,13 +331,13 @@ func TestAlterDBPlacement(t *testing.T) {
 	// Test for Alter Default Placement Policy, And will not update the old table options.
 	tk.MustExec("ALTER DATABASE TestAlterDB DEFAULT PLACEMENT POLICY=`alter_y`;")
 	tk.MustExec("create table t2(a int);")
-	tk.MustQuery(`show create table t`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create table t`).Check(testkit.RowsWithSep("|",
 		"t CREATE TABLE `t` (\n"+
 			"  `a` int(11) DEFAULT NULL\n"+
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin "+
 			"/*T![placement] PLACEMENT POLICY=`alter_x` */",
 	))
-	tk.MustQuery(`show create table t2`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create table t2`).Check(testkit.RowsWithSep("|",
 		"t2 CREATE TABLE `t2` (\n"+
 			"  `a` int(11) DEFAULT NULL\n"+
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin "+
@@ -353,7 +352,7 @@ func TestAlterDBPlacement(t *testing.T) {
 
 	// Test for Alter Placement Rule affect table created.
 	tk.MustExec("create table t3(a int);")
-	tk.MustQuery(`show create table t3`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create table t3`).Check(testkit.RowsWithSep("|",
 		"t3 CREATE TABLE `t3` (\n"+
 			"  `a` int(11) DEFAULT NULL\n"+
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin "+
@@ -363,7 +362,7 @@ func TestAlterDBPlacement(t *testing.T) {
 
 	// Test for override default option
 	tk.MustExec("create table t4(a int) PLACEMENT POLICY=\"alter_y\";")
-	tk.MustQuery(`show create table t4`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create table t4`).Check(testkit.RowsWithSep("|",
 		"t4 CREATE TABLE `t4` (\n"+
 			"  `a` int(11) DEFAULT NULL\n"+
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin "+
@@ -373,7 +372,7 @@ func TestAlterDBPlacement(t *testing.T) {
 
 	// Test alter to another policy
 	tk.MustExec("ALTER DATABASE TestAlterDB PLACEMENT POLICY=`alter_y`;")
-	tk.MustQuery(`show create database TestAlterDB`).Check(testutil.RowsWithSep("|",
+	tk.MustQuery(`show create database TestAlterDB`).Check(testkit.RowsWithSep("|",
 		"TestAlterDB CREATE DATABASE `TestAlterDB` /*!40100 DEFAULT CHARACTER SET utf8mb4 */ "+
 			"/*T![placement] PLACEMENT POLICY=`alter_y` */",
 	))
